requirements.txt
AdvancedHTTPServer==2.2.0
ajpy==0.0.4
alembic==1.0.11.dev0
aniso8601==7.0.0
asn1crypto==0.24.0
basemap==1.2.0
bcrypt==3.1.6
beautifulsoup4==4.8.0
binwalk==2.1.2
blinker==1.4
boltons==19.1.0
Brlapi==0.6.7
Brotli==1.0.7
certifi==2018.8.24
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.3.7
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
debtags==2.1
decorator==4.3.0
dicttoxml==1.7.4
distro==1.3.0
distro-info==0.21
Django==2.2.5
django-bootstrap3-datetimepicker==2.2.3
djangorestframework==3.10.3
dnspython==1.16.0
ecdsa==0.13.2
email-validator==1.0.3
entrypoints==0.3
Flask==1.0.2
future==0.16.0
geoip2==2.9.0
geojson==2.5.0
graphene==2.1.7
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
h11==0.8.1
h2==3.0.1
hashID==3.1.4
hpack==3.0.0
html5lib==1.0.1
httplib2==0.11.3
hyperframe==5.2.0
icalendar==4.0.3
idna==2.6
IPy==0.83
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10.1
jsonschema==2.6.0
jupyter-core==4.5.0
kaitaistruct==0.8
keyring==17.1.1
keyrings.alt==3.1.1
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2019.7.0
KismetCaptureRtladsb==2018.7.1
KismetCaptureRtlamr==2018.7.0
kiwisolver==1.0.1
ldap3==2.5.1
louis==3.10.0
lxml==4.3.3
mailer==0.8.1
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.0.2
maxminddb==1.4.1
mitmproxy==4.0.4
msgpack==0.5.6
mysqlclient==1.3.10
nassl==2.1.2
nbformat==4.4.0
netaddr==0.7.19
numpy==1.16.2
olefile==0.46
openvas-lib==1.1.4
packaging==19.0
paramiko==2.6.0
passlib==1.7.1
patator==0.7
Pillow==6.1.0
plotly==4.1.0
pluginbase==1.0.0
ply==3.11
promise==2.2
protobuf==3.6.1
psycopg2==2.7.7
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.32.2
pyinotify==0.9.6
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyotp==2.3.0
pyparsing==2.2.0
pyperclip==1.6.4
pyproj==1.9.6
pyqtgraph==0.10.0
pyserial==3.4
pyshp==2.1.0
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
pysmi==0.3.2
pysnmp==4.4.6
python-apt==1.8.4
python-dateutil==2.7.3
python-debian==0.1.35
python-debianbts==2.8.2
python-editor==1.0.3
python-pam==1.8.4
pytz==2019.2
pyxdg==0.25
PyYAML==3.13
reportbug==7.5.2
requests==2.21.0
requests-file==1.4.3
retrying==1.3.3
ruamel.yaml==0.15.34
rule-engine==1.1.0
Rx==1.6.1
scipy==1.2.2
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.18
six==1.12.0
smoke-zephyr==1.4.1
sortedcontainers==2.1.0
soupsieve==1.9.2
SQLAlchemy==1.3.1
sqlparse==0.3.0
sslyze==2.0.6
termcolor==1.1.0
tls-parser==1.2.1
tornado==5.1.1
traitlets==4.3.2
tzlocal==1.5.1
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.24.1
urwid==2.0.1
wafw00f==1.0.0
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.14.1
wfuzz==2.4
wsproto==0.11.0
XlsxWriter==1.1.2
yara-python==3.10.0
